Temporary video
I made a video for my class... that is way too long (40 minutes) and way to boring. All the same it goes through the inputs, explains what they are, how to get them (not prep them), and how to run GUT. It also goes through outputs. It should be broken into smaller pieces for each thing. But here you go for something.